Create README.md
Project Description (Brief Overview)
The objective of this project is to identify and analyze the primary and secondary reasons behind customer withdrawal from a telecommunications company, with a focus on geographical trends and demographic patterns. The analysis aims to uncover key insights into customer dissatisfaction. This will support the development of strategic recommendations to reduce churn, improve customer retention, and enhance overall service quality in targeted regions.